wip: hacky pure Go report
