+++ src/fixtures/func
@@ -1,0 +1,3 @@
+bar.go:114:		Bar				67.0%
+qux.go:114:		Qux				78.4%
+total:									(statements)				48.7%
