--- src/fixtures/test-no-branch.xml
@@ -1,70 +1,0 @@
-<?xml version="1.0"?>
-<!--DOCTYPE coverage SYSTEM "http://cobertura.sourceforge.net/xml/coverage-03.dtd"-->
-
-<coverage line-rate="0.9" branch-rate="0" version="1.9" timestamp="1187350905008">
-  <sources>
-    <source>C:/local/mvn-coverage-example/src/main/java</source>
-    <source>--source</source>
-  </sources>
-  <packages>
-    <package name="" line-rate="1.0" branch-rate="0" complexity="0">
-      <classes>
-        <class name="Main" filename="Main.java" line-rate="1.0" branch-rate="0" complexity="0">
-          <lines>
-            <line number="10" hits="3"/>
-            <line number="16" hits="3"/>
-            <line number="17" hits="3"/>
-            <line number="18" hits="3"/>
-            <line number="19" hits="3"/>
-            <line number="23" hits="3"/>
-            <line number="25" hits="3"/>
-            <line number="26" hits="3"/>
-            <line number="28" hits="3"/>
-            <line number="29" hits="3"/>
-            <line number="30" hits="3"/>
-          </lines>
-        </class>
-      </classes>
-    </package>
-    <package name="search" line-rate="0.8421052631578947" branch-rate="0" complexity="3.25">
-      <classes>
-        <class name="search.BinarySearch" filename="search/BinarySearch.java" line-rate="0.9166666666666666"
-               branch-rate="0" complexity="0">
-          <lines>
-            <line number="12" hits="3"/>
-            <line number="16" hits="3"/>
-            <line number="18" hits="12" />
-            <line number="20" hits="9"/>
-            <line number="21" hits="9"/>
-            <line number="23" hits="9"/>
-            <line number="24" hits="0"/>
-            <line number="25" hits="9"/>
-            <line number="26" hits="6"/>
-            <line number="28" hits="3"/>
-            <line number="29" hits="9"/>
-            <line number="31" hits="3"/>
-          </lines>
-        </class>
-        <class name="search.ISortedArraySearch" filename="search/ISortedArraySearch.java" line-rate="1.0"
-               branch-rate="0" complexity="0">
-          <methods>
-          </methods>
-          <lines>
-          </lines>
-        </class>
-        <class name="search.LinearSearch" filename="search/LinearSearch.java" line-rate="0.7142857142857143"
-               branch-rate="0" complexity="0.0">
-          <lines>
-            <line number="9" hits="3"/>
-            <line number="13" hits="9"/>
-            <line number="15" hits="9"/>
-            <line number="16" hits="3"/>
-            <line number="17" hits="6"/>
-            <line number="19" hits="0"/>
-            <line number="24" hits="0"/>
-          </lines>
-        </class>
-      </classes>
-    </package>
-  </packages>
-</coverage>
