--- src/fixtures/test-python.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" ?>
-<coverage branch-rate="0" line-rate="0.9" timestamp="1460488624050" version="4.0.3">
-  <!-- Generated by coverage.py: https://coverage.readthedocs.org -->
-  <!-- Based on https://raw.githubusercontent.com/cobertura/web/f0366e5e2cf18f111cbd61fc34ef720a6584ba02/htdocs/xml/coverage-03.dtd -->
-  <sources>
-    <source></source>
-  </sources>
-  <packages>
-    <package branch-rate="0" complexity="0" line-rate="0.9" name=".">
-      <classes>
-        <class branch-rate="0" complexity="0" filename="source.py" line-rate="0.9" name="source.py">
-          <methods/>
-          <lines>
-            <line hits="1" number="4"/>
-            <line hits="1" number="5"/>
-            <line hits="1" number="6"/>
-            <line hits="1" number="7"/>
-            <line hits="1" number="9"/>
-            <line hits="1" number="10"/>
-            <line hits="1" number="12"/>
-            <line hits="0" number="13"/>
-            <line hits="1" number="15"/>
-            <line hits="1" number="16"/>
-          </lines>
-        </class>
-      </classes>
-    </package>
-  </packages>
-</coverage>
