+++ src/gocover.js
@@ -1,0 +1,131 @@
+const events = require("events");
+const fs = require("fs").promises;
+const glob = require("glob-promise");
+const readline = require("readline");
+
+/**
+ * generate the report for the given file
+ *
+ * @param path: string
+ * @param options: object
+ * @return {Promise<{total: number, files: T[]}>}
+ */
+async function readCoverageFromFile(path, options) {
+  const file = await fs.open(path);
+  const fileStream = file.createReadStream(path);
+  const rl = readline.createInterface({
+    input: fileStream,
+    crlfDelay: Infinity,
+  });
+
+  var total;
+  const files = [];
+
+  for await (const line of rl) {
+    const [result, filename, time, coverage] = line.split("\t");
+    switch (result.trim()) {
+      case 'ok':
+        files.push({
+          filename,
+          total: pct(coverage)
+        })
+        break;
+      case '?':
+          files.push({
+            filename,
+            total: 0.0
+          });
+        break;
+      case 'total:':
+        total = pct(line);
+        break;
+    }
+  }
+  /* const xml = await fs.readFile(path, "utf-8");
+  const { coverage } = await parseString(xml, {
+    explicitArray: false,
+    mergeAttrs: true,
+  });
+  const { packages } = coverage;
+  const classes = processPackages(packages);
+  const files = classes
+    .filter(Boolean)
+    .map((klass) => {
+      return {
+        ...calculateRates(klass),
+        filename: klass["filename"],
+        name: klass["name"],
+        missing: missingLines(klass),
+      };
+    })
+    .filter((file) => options.skipCovered === false || file.total < 100);
+  return {
+    ...calculateRates(coverage),
+    files,
+  }; */
+  return { total, files };
+}
+
+const pct = str => parseFloat(str.match(/([0-9]+.[0-9]+)%/)[1], 10)
+
+/**
+ *
+ * @param path: string
+ * @param options: {}
+ * @returns {Promise<{total: number, folder: string, files: T[]}[]>}
+ */
+async function processCoverage(path, options) {
+  options = options || { skipCovered: false };
+
+  const paths = glob.hasMagic(path) ? await glob(path) : [path];
+  const positionOfFirstDiff = longestCommonPrefix(paths);
+  return await Promise.all(
+    paths.map(async (path) => {
+      const report = await readCoverageFromFile(path, options);
+      const folder = trimFolder(path, positionOfFirstDiff);
+      return {
+        ...report,
+        folder,
+      };
+    })
+  );
+}
+
+function trimFolder(path, positionOfFirstDiff) {
+  const lastFolder = path.lastIndexOf("/") + 1;
+  if (positionOfFirstDiff >= lastFolder) {
+    return path.substr(lastFolder);
+  } else {
+    const startOffset = Math.min(positionOfFirstDiff - 1, lastFolder);
+    const length = path.length - startOffset - lastFolder - 2; // remove filename
+    return path.substr(startOffset, length);
+  }
+}
+
+/**
+ *
+ * @param paths: [string]
+ * @returns number
+ */
+function longestCommonPrefix(paths) {
+  let prefix = "";
+  if (paths === null || paths.length === 0) return 0;
+
+  for (let i = 0; i < paths[0].length; i++) {
+    const char = paths[0][i]; // loop through all characters of the very first string.
+
+    for (let j = 1; j < paths.length; j++) {
+      // loop through all other strings in the array
+      if (paths[j][i] !== char) return prefix.length;
+    }
+    prefix = prefix + char;
+  }
+
+  return prefix.length;
+}
+
+module.exports = {
+  processCoverage,
+  longestCommonPrefix,
+  trimFolder,
+};
